Feat/카드 섞기

+++ src/screens/SettingScreen.tsx
@@ -1,29 +1,45 @@
-import {StyleSheet, Text, View} from 'react-native';
-import React, {useState} from 'react';
-import {useDispatch, useSelector} from 'react-redux';
+import React from 'react';
+import {View, StyleSheet, Alert} from 'react-native';
 import {Button} from 'react-native-paper';
+import {useDispatch} from 'react-redux';
 
 import {resetFavorite} from '../redux/slices/favoriteQuoteSlice';
 import {removeAllStar} from '../redux/slices/allQuoteSlice';
 
-type Props = {};
-
-const SettingScreen = (props: Props) => {
+const SettingScreen = () => {
   const dispatch = useDispatch();
 
-  return (
-    <View>
-      <Button
-        onPress={() => {
+  const removeFavoriteHandler = () => {
+    Alert.alert('Warning!!', 'You really want to delete all favorite lists?', [
+      {
+        text: 'CANCEL',
+        onPress: () => {
+          return;
+        },
+      },
+      {
+        text: 'DELETE',
+        onPress: () => {
           dispatch(resetFavorite());
           dispatch(removeAllStar());
-        }}>
-        Reset ALL Favorite
-      </Button>
+        },
+      },
+    ]);
+  };
+
+  return (
+    <View style={styles.container}>
+      <Button onPress={removeFavoriteHandler}>Reset ALL Favorite</Button>
     </View>
   );
 };
 
-export default SettingScreen;
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
 
-const styles = StyleSheet.create({});
+export default SettingScreen;
